requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.3
Flask-Assets==0.8
Flask-Script==0.6.3
Jinja2==2.7.1
MarkupSafe==0.18
SQLAlchemy==0.8.3
WTForms==1.0.5
Werkzeug==0.9.4
itsdangerous==0.23
webassets==0.9
pyScss==1.2.0.post3
six==1.4.1
PyYAML==3.10
requests==2.1.0
alembic==0.6.1
Mako==0.9.0
Flask-Migrate==1.0.0
